search by title api route completed

+++ controllers/videoController.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Video from "../models/VideoModel.js";
 
 // get all videos
@@ -9,15 +10,22 @@ const getVideos = async (req, res) => {
 
 // get video by id
 const getVideoById = async (req, res) => {
-    const video = await Video.findById(req.params.videoId);
+    
+    if(mongoose.Types.ObjectId.isValid(req.params.videoId)) {
+        const video = await Video.findById(req.params.videoId);
 
-    if(video) {
-        res.json(video)
-    } else {
-        res.status(404).json({ message: 'video with this id not found!'})
-        throw new Error('Video Not Found')
+        if(video) {
+            res.json(video)
+        } else {
+            res.status(404).json({ message: 'video with this id not found!'})
+            throw new Error('Video Not Found')
+        }
     }
+
+    
 }
+
+
 
 
 // add video
@@ -46,46 +54,54 @@ const createVideo = async (req, res) => {
 const updateVideo = async (req, res) => {
 
     const {title, category, description, thumbnailUrl, videoUrl } = req.body
-
-    const video = await Video.findById(req.params.videoId);
-
-    if(video) {
-        video.title = title
-        video.category = category
-        video.description = description
-        video.thumbnailUrl = thumbnailUrl
-        video.videoUrl = videoUrl
-
-        const updateVideo = await video.save()
-
-        res.json(updateVideo);
+   
+    if(mongoose.Types.ObjectId.isValid(req.params.videoId)) {
+        const video = await Video.findById(req.params.videoId);
+        if(video) {
+            video.title = title
+            video.category = category
+            video.description = description
+            video.thumbnailUrl = thumbnailUrl
+            video.videoUrl = videoUrl
+    
+            const updateVideo = await video.save()
+    
+            res.json(updateVideo);
+        }
+         else {
+            res.status(404)
+            throw new Error('Video Not Found')
+        }
     }
-     else {
-        res.status(404)
-        throw new Error('Video Not Found')
-    }
+    
+
+   
        
 }
 
-// delete video by id
+//delete video by id
 const deleteVideoById = async (req, res) => {
-    const foundedVideo =  await Video.findById(req.params.videoId);
+    if(mongoose.Types.ObjectId.isValid(req.params.videoId)) {
+        const foundedVideo =  await Video.findById(req.params.videoId);
 
-    if(foundedVideo) {
-
-     await Video.deleteOne(foundedVideo)
-     res.json({ message: 'Video Deleted' })
-        
-    } else {
-        res.status(404)
-        throw new Error('Video Not Found')
+        if(foundedVideo) {
+    
+         await Video.deleteOne(foundedVideo)
+         res.json({ message: 'Video Deleted' })
+            
+        } else {
+            res.status(404)
+            throw new Error('Video Not Found')
+        }
     }
+   
     
 }
 
 // update like count
 const updateLikeCount = async (req, res) => {
 
+    if(mongoose.Types.ObjectId.isValid(req.params.videoId)) {
         const videoId = await Video.findById(req.params.videoId);
 
         if(videoId) {
@@ -95,6 +111,8 @@ const updateLikeCount = async (req, res) => {
             res.status(404)
             throw new Error('Not Found any video to like count')
         }
+    }
+        
        
 
     // if(foundedVideo) {
@@ -108,4 +126,26 @@ const updateLikeCount = async (req, res) => {
     // }
     
 }
-export { getVideos, getVideoById, createVideo, updateVideo, deleteVideoById, updateLikeCount }
+
+// search 
+const searchByTitle = async (req, res) => {
+    const queryTitle = new RegExp(req.params.title, 'i');
+    if(queryTitle !== '') {
+        try {
+            const search_results =  await Video.find({ title: queryTitle })
+            res.status(200).json(search_results)
+        } catch (error) {
+            console.log(error)
+        }
+       
+    } else {
+        res.status(404)
+        throw new Error('Not Found based on this title')
+    }
+   
+}
+
+
+
+
+export { getVideos, getVideoById, createVideo, updateVideo, deleteVideoById, updateLikeCount, searchByTitle }
